Assorted small fixes and improvements
- Improve accept header message in bulk_data_export
- Add prefer_header? method
- Error when _type or _outputFormat are passed as params
- More strictly validate _since parameter
- Change params.permit to params.require when applicable
- Server error when transaction_time is not a date in export_status
- Change :filename in export_files url to :resource_type
- Add 'Expires' header in export_status
- Check 'Accept' header in export_files
- Add line to purge job to purge ApiDownloads over 1 day old
- Refactor tests in controllers/fhir/r4 to inherit from one
  ApiControllerTestCase class
