+++ test/controllers/fhir/r4/immunization_test.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
-require 'test_helper'
-require 'rspec/mocks/minitest_integration'
-require 'controllers/fhir/r4/api_controller_test'
+require 'api_controller_test_case'
 
-class ApiControllerTest < ActionDispatch::IntegrationTest
+class ImmunizationTest < ApiControllerTestCase
   setup do
+    setup_system_applications
+    setup_system_tokens
+    setup_user_applications
     setup_vaccines
+    setup_logger
   end
 
   def setup_vaccines
